Cleanup and expand wheel job definitions
This commit makes a few small changes to the cibuildwheel ci job
definitions. The first change it makes is to unify the job configuration
to a single location in the pyproject.toml. In recent versions of
cibuildwheel you can set the configuration in the pyproject.toml (or
optionally a different toml file) to have a unified config for all jobs.
This makes it simpler to maintain as we can have a common config and
then per job overrides as needed. As part of this several rough edges in
the config are changed mainly around supported environments so we only
build in environments that qiskit currently supports. The second change
here is to expand the build matrix to enable cross-builds for other CPU
architectures. This is mainly arm64/aarch64 on Linux and macOS, pp64le
on linux, and s390x on linux. Since we don't have native hardware
available to compile on these platforms in CI we either leverage cross
compilation (on macos arm64) or qemu emulation. However as emulation is
painfully slow, or in the case of cross builds the binaries are
untestable these jobs are reserved just for release time as we don't
want this overhead during normal commits.