Simplify instruction duration to cycles.
The previous approach used a linear interpolation, which wouldn't always preserve
relative durations and didn't guarantee that instructions with different durations
wouldn't collide when converted.

This approach doesn't guarantee that either, but it's a lot simpler. With this
approach, the fastest non-zero duration (min_duration) instruction is considered
to take 1 cycle to execute. For other instructions,
cycles = ceil(duration/min_duration).