Add ToqmProfile for configuration settings.
There are three default profiles (ToqmProfileO1, ToqmProfileO2, ToqmProfileO3)
which can be used to change the configuration used to invoke libtoqm.

No real rigorous testing was done in determining these settings,
so they're likely to change.

The idea is that Terra should be able to point to these profiles,
allowing us to tweak settings between releases.