Add ToqmStrategy for configuration settings. (#10)
Factors the configuration and invocation of libtoqm out into ToqmStrategy, which can be subclassed to customize the run based on the circuit and target device.

There are three default strategies (ToqmStrategyO1, ToqmStrategyO2, ToqmStrategyO3) corresponding to optimization level which can be used to change the configuration used to invoke libtoqm.

No real rigorous testing was done in determining these settings, so they're likely to change.

The idea is that Terra should be able to point to these optimization level-based strategies, allowing us to tweak settings between releases.