Use Latency126 for heuristic mappers.
Using Table latency seems to cause nontermination for some circuits. Currently,
it's thought that this only happens for heuristic mappers, but it's possible
the issue is present for optimal mappers as well. I just haven't found such a
case yet. Ultimately, it seems like there's a preexisting bug in TOQM that
gets exposed when we consider coupling-specific latencies.

Settings for optimization levels have also been tweaked so that mappings
terminate faster. These settings are based on some preliminary
benchmarks for QV256 circuits, and will be updated further once a more
comprehensive analysis of those results is performed.

The normalize factor has also been increase in ToqmSwap from 1 to 2 so
that Table durations have more resolution (i.e. gate durations that are
close should less often round to the same integer).