Remove layout option and tweak settings.
The layout option really isn't useful since it isn't compatible with heuristic
mappers (anything involving GreedyMapper), and in fact breaks things when
enabled for these configurations. Therefore, we just remove it, and enable
layout changes always for optimal mappers and leave it unspecified (which
behaves as enabled) for heuristic mappers.